Basic user registration workflow.
Users go to registration page and enter email address. They are then sent an
email (not implemented yet) with a URL that contains a token. The URL leads
them to a form where they can enter a password and location. The latitude and
longitude coordinates are calculated based on the location using the Google
Maps API. The user is then marked as 'confirmed' in the database and redirected
to the home page.